move ApplyTrustedCorrelationHeaders into the shared HttpTaskInvocation core
